requirements.txt
pandas==2.0.3
sentence_transformers==3.0.1
tabulate==0.9.0
openai==1.76.0
unsloth==2025.4.4
numpy==1.26.0
langchain
langchain-community
langchain-core
hf_xet
scikit-learn==1.6.1
tensorboardX==2.6.2.2
pytorch-transformers==1.2.0